backend/requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
arxiv==2.2.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
distro==1.9.0
etelemetry==0.3.1
fastapi==0.116.1
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-ollama==0.3.6
langchain-text-splitters==0.3.9
langsmith==0.4.14
looseversion==1.3.0
lxml==6.0.0
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
neo4j==5.28.2
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.2
ollama==0.5.3
openai==1.99.9
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
propcache==0.3.2
prov==2.1.1
puremagic==1.30
py4j==0.10.9.9
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
PyMuPDF==1.26.3
pyparsing==3.2.3
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
scipy==1.16.1
sgmllib3k==1.0.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
traits==7.0.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
